Emails: Move BillingIntervalToggle component to new folder
